requirements.txt
async-generator==1.10
attrs==22.1.0
boto==2.49.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==108.0.5359.22.0
click==8.1.3
cmake==3.24.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
dlib==19.24.0
exceptiongroup==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
fonttools==4.38.0
future==0.18.2
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
MouseInfo==0.1.3
numpy==1.23.4
opencv-python==4.6.0.66
openpifpaf==0.13.7
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyocr==0.8.3
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pysparkling==0.6.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytweening==1.0.4
pytz==2022.6
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
torch==1.13.0
torchvision==0.14.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.10.0
